requirements.txt
async-timeout==5.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
exceptiongroup==1.3.0
idna==3.11
iniconfig==2.1.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
six==1.17.0
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
